Add 'keeping the context' examples
Either adding a review to a restaurant or removing a review do not
cause a scroll to top.
Forms and links can be identified by classes as anchors. If so,
they would have an event listener that would the save the scroll x
position when clicked / submitted.
that position would be saved to `scrollTo` on load.